+++ MQTTPacket/src/MQTTDeserializePublish.c
@@ -50,7 +50,7 @@ int32_t MQTTDeserialize_publish(unsigned char* dup, unsigned char* qos, unsigned
 	unsigned char* curdata = buf;
 	const unsigned char* enddata = NULL;
 	int32_t rc = 0;
-	size_t mylen = 0;
+	uint32_t mylen = 0;
 
 	FUNC_ENTRY;
 	header = readChar(&curdata);
@@ -107,7 +107,7 @@ int32_t MQTTDeserialize_ack(unsigned char* packettype, unsigned char* dup, unsig
 	unsigned char* curdata = buf;
 	unsigned char* enddata = NULL;
 	int32_t rc = 0;
-	size_t mylen = 0;
+	uint32_t mylen = 0;
 
 	FUNC_ENTRY;
 	header = readChar(&curdata);
